# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Agentic AI career assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CareerProfile {
  id: number;
  name: string;
  headline: string;
  education: string;
  skills: string[];
  interests: string[];
  goals: string[];
  targetRoles: string[];
  updatedAt: string;
}

export interface CareerProfileInput {
  name: string;
  headline: string;
  education: string;
  skills: string[];
  interests: string[];
  goals: string[];
  targetRoles: string[];
}

export interface CareerRole {
  title: string;
  fit: number;
  reason: string;
  nextStep: string;
}

export interface SkillGap {
  skill: string;
  priority: string;
  rationale: string;
  action: string;
}

export interface CareerActivity {
  id: number;
  type: string;
  title: string;
  detail: string;
  createdAt: string;
}

export interface CareerDashboard {
  profile: CareerProfile;
  resumeScore: number;
  skillCoverage: number;
  indexedDocuments: number;
  interviewReadiness: number;
  topRoles: CareerRole[];
  skillGaps: SkillGap[];
  recentActivity: CareerActivity[];
}

export interface ResumeAnalysisInput {
  filename: string;
  resumeText: string;
  targetRole?: string;
}

export interface ResumeAnalysis {
  id: number;
  filename: string;
  score: number;
  summary: string;
  strengths: string[];
  improvements: string[];
  missingSkills: SkillGap[];
  roleMatches: CareerRole[];
  agentTrace: string[];
  createdAt: string;
}

export interface CareerDocumentInput {
  filename: string;
  content: string;
  kind: string;
}

export interface CareerDocument {
  id: number;
  filename: string;
  kind: string;
  excerpt: string;
  chunks: number;
  createdAt: string;
}

export interface CareerQuestionInput {
  question: string;
}

export interface CareerAnswer {
  question: string;
  answer: string;
  sources: string[];
  agentTrace: string[];
}

export interface ResumeToolInput {
  targetRole: string;
  focus: string;
}

export interface ResumeToolOutput {
  title: string;
  content: string;
  highlights: string[];
  toolTrace: string[];
}

export interface SkillToolInput {
  targetRole: string;
  timeframe: string;
}

export interface SkillPlanItem {
  week: string;
  skill: string;
  why: string;
  project: string;
}

export interface SkillToolOutput {
  targetRole: string;
  plan: SkillPlanItem[];
  toolTrace: string[];
}

export interface InterviewToolInput {
  targetRole: string;
  difficulty: string;
}

export interface InterviewQuestion {
  category: string;
  question: string;
  coaching: string;
}

export interface InterviewToolOutput {
  targetRole: string;
  questions: InterviewQuestion[];
  toolTrace: string[];
}