# lib/api-zod/src/generated/types/careerAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Agentic AI career assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerAnswer {
  question: string;
  answer: string;
  sources: string[];
  agentTrace: string[];
}