# lib/api-zod/src/generated/types/careerDocument.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Agentic AI career assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerDocument {
  id: number;
  filename: string;
  kind: string;
  excerpt: string;
  chunks: number;
  createdAt: string;
}